[#254]  Refactor installdependencies MOJO to decorator inspired pattern
