+++ habushu-maven-plugin/pom.xml
@@ -196,6 +196,12 @@
             <artifactId>junit-platform-console</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-text</artifactId>
+            <version>1.12.0</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
