+++ habushu-maven-plugin/src/main/java/org/technologybrewery/habushu/InstallDependenciesPoetry.java
@@ -3,10 +3,8 @@ package org.technologybrewery.habushu;
 import com.electronwill.nightconfig.core.CommentedConfig;
 import com.electronwill.nightconfig.core.Config;
 import com.electronwill.nightconfig.core.file.FileConfig;
-import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.collections4.MapUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.http.client.utils.URIBuilder;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugin.logging.Log;
@@ -18,11 +16,11 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
-import java.net.URISyntaxException;
-import java.nio.file.Files;
-import java.nio.file.StandardOpenOption;
-import java.time.LocalDateTime;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -37,8 +35,6 @@ import java.util.concurrent.TimeUnit;
 public class InstallDependenciesPoetry extends AbstractInstallDependencies {
 
     private static final String EQUALS = "=";
-    private static final String SNAPSHOT = "-SNAPSHOT";
-    protected static final String PUBLIC_PYPI_REPO_ID = "pypi";
     private static final String POETRY_CLEAN_CACHE_COMMAND = "poetry cache clear . --all";
 
     /**
@@ -48,8 +44,8 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
      */
     protected static final String PYPROJECT_PACKAGE_SOURCES_PATH = "tool.poetry.source";
 
-    public InstallDependenciesPoetry(File baseDir, Log log, InstallDependenciesConfigurations installDependenciesConfigurations)  {
-        super(baseDir, log, installDependenciesConfigurations);
+    public InstallDependenciesPoetry(File baseDir, Log log, InstallDependenciesMojo mojo)  {
+        super(baseDir, log, mojo);
     }
 
 
@@ -61,15 +57,15 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
 
         processManagedDependencyMismatches();
 
-        prepareRepositoryForInstallation(installDependenciesConfigurations.getPypiRepoId(), installDependenciesConfigurations.getPypiRepoUrl());
-        if (installDependenciesConfigurations.useDevRepository()) {
-            prepareRepositoryForInstallation(installDependenciesConfigurations.getDevRepositoryId(), installDependenciesConfigurations.getDevRepositoryUrl());
+        prepareRepositoryForInstallation(installDependenciesMojo.getPypiRepoId(), installDependenciesMojo.getPypiRepoUrl(), PYPROJECT_PACKAGE_SOURCES_PATH);
+        if (installDependenciesMojo.useDevRepository()) {
+            prepareRepositoryForInstallation(installDependenciesMojo.getDevRepositoryId(), installDependenciesMojo.getDevRepositoryUrl(), PYPROJECT_PACKAGE_SOURCES_PATH);
         }
 
-        if (!installDependenciesConfigurations.skipPoetryLockUpdate()) {
+        if (!installDependenciesMojo.skipPoetryLockUpdate()) {
             log.info("Locking dependencies specified in pyproject.toml...");
             poetryHelper.executePackageManagerCommandAndLogAfterTimeout(
-                    poetryHelper.createLockCommand(installDependenciesConfigurations.skipPoetryLockUpdate()),
+                    poetryHelper.createLockCommand(installDependenciesMojo.skipPoetryLockUpdate()),
                     2,
                     TimeUnit.MINUTES,
                     POETRY_CLEAN_CACHE_COMMAND
@@ -77,13 +73,13 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
 
         }
 
-        List<String> installCommand = poetryHelper.createInstallCommand(installDependenciesConfigurations.isForceSync());
+        List<String> installCommand = poetryHelper.createInstallCommand(installDependenciesMojo.forceSync());
 
-        for (String groupName : installDependenciesConfigurations.getWithGroups()) {
+        for (String groupName : installDependenciesMojo.getWithGroups()) {
             installCommand.add("--with");
             installCommand.add(groupName);
         }
-        for (String groupName : installDependenciesConfigurations.getWithoutGroups()) {
+        for (String groupName : installDependenciesMojo.getWithoutGroups()) {
             installCommand.add("--without");
             installCommand.add(groupName);
         }
@@ -100,10 +96,10 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
         String currentInProjectSetting = poetryHelper.execute(arguments);
         
         // update the poetry config to match the useInProjectVirtualEnvironment boolean
-        if (installDependenciesConfigurations.useInProjectVirtualEnvironment() && Boolean.FALSE.equals(Boolean.valueOf(currentInProjectSetting))) {
+        if (installDependenciesMojo.useInProjectVirtualEnvironment() && Boolean.FALSE.equals(Boolean.valueOf(currentInProjectSetting))) {
             log.info("Configuring Poetry to use an in-project virtual environment...");
             configureVirtualEnvironmentsInProject(true);
-        } else if (!installDependenciesConfigurations.useInProjectVirtualEnvironment() && Boolean.TRUE.equals(Boolean.valueOf(currentInProjectSetting))) {
+        } else if (!installDependenciesMojo.useInProjectVirtualEnvironment() && Boolean.TRUE.equals(Boolean.valueOf(currentInProjectSetting))) {
             configureVirtualEnvironmentsInProject(false);
         }
     }
@@ -118,102 +114,8 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
         poetryHelper.executeAndLogOutput(arguments);
     }
 
-    private void prepareRepositoryForInstallation(String repoId, String repoUrl) throws MojoExecutionException {
-        if (StringUtils.isNotEmpty(repoUrl) && installDependenciesConfigurations.addPypiRepoAsPackageSources()) {
-            String pypiRepoSimpleIndexUrl;
-            try {
-                pypiRepoSimpleIndexUrl = getPyPiRepoSimpleIndexUrl(repoUrl);
-            } catch (URISyntaxException e) {
-                throw new MojoExecutionException(
-                        String.format("Could not parse configured repoUrl %s", repoUrl), e);
-            }
-
-            // NB later version of Poetry will support retrieving and configuring package
-            // source repositories via the "poetry source" command in future releases, but
-            // for now we need to manually inspect and modify the package's pyproject.toml
-            Config matchingPypiRepoSourceConfig;
-            try (FileConfig pyProjectConfig = FileConfig.of(getPyProjectTomlFile())) {
-                pyProjectConfig.load();
-
-                Optional<List<Config>> packageSources = pyProjectConfig.getOptional(PYPROJECT_PACKAGE_SOURCES_PATH);
-                matchingPypiRepoSourceConfig = packageSources.orElse(Collections.emptyList()).stream()
-                        .filter(packageSource -> pypiRepoSimpleIndexUrl.equals(packageSource.get("url"))).findFirst()
-                        .orElse(Config.inMemory());
-            }
-
-            if (!matchingPypiRepoSourceConfig.isEmpty()) {
-                if (log.isDebugEnabled()) {
-                    log.debug(String.format(
-                            "Configured PyPi repository %s found in the following pyproject.toml [[%s]] array element: %s",
-                            repoUrl, PYPROJECT_PACKAGE_SOURCES_PATH, matchingPypiRepoSourceConfig));
-                }
-            } else {
-                // NB NightConfig's TOML serializer generates TOML in a manner that makes it
-                // difficult to append an array element of tables to an existing TOML
-                // configuration, so manually write out the desired new repository TOML
-                // configuration with human-readable formatting
-                List<String> newPypiRepoSourceConfig = Arrays.asList(System.lineSeparator(), String.format(
-                                "# Added by habushu-maven-plugin at %s to use %s as source PyPi repository for installing dependencies",
-                                LocalDateTime.now(), pypiRepoSimpleIndexUrl),
-                        String.format("[[%s]]", PYPROJECT_PACKAGE_SOURCES_PATH),
-                        String.format("name = \"%s\"",
-                                StringUtils.isNotEmpty(repoId) && !PUBLIC_PYPI_REPO_ID.equals(repoId)
-                                        ? repoId
-                                        : "private-pypi-repo"),
-                        String.format("url = \"%s\"", pypiRepoSimpleIndexUrl), "priority = \"supplemental\"");
-                log.info(String.format("Private PyPi repository entry for %s not found in pyproject.toml",
-                        repoUrl));
-                log.info(String.format(
-                        "Adding %s to pyproject.toml as supplemental repository from which dependencies may be installed",
-                        pypiRepoSimpleIndexUrl));
-                try {
-                    Files.write(getPyProjectTomlFile().toPath(), newPypiRepoSourceConfig,
-                            StandardOpenOption.APPEND);
-                } catch (IOException e) {
-                    throw new MojoExecutionException(String.format(
-                            "Could not write new [[%s]] element to pyproject.toml", PYPROJECT_PACKAGE_SOURCES_PATH), e);
-                }
-            }
-
-        }
-    }
-
-    /**
-     * Attempts to infer the PEP-503 compliant PyPI simple repository index URL
-     * associated with the provided PyPI repository URL. In order to configure
-     * Poetry to use a private PyPi repository as a source for installing package
-     * dependencies, the simple index URL of the repository <b>*must*</b> be
-     * utilized. For example, if a private PyPI repository is hosted at
-     * https://my-company-sonatype-nexus/repository/internal-pypi and provided to
-     * Habushu via the {@literal <pypiRepoUrl>} configuration, the simple index URL
-     * returned by this method will be
-     * https://my-company-sonatype-nexus/repository/internal-pypi/simple/ (the
-     * trailing slash is required!).
-     *
-     * @param pypiRepoUrl URL of the private PyPi repository for which to generate
-     *                    the simple index API URL.
-     * @return simple index API URL associated with the given PyPi repository URL.
-     * @throws URISyntaxException
-     */
-    protected String getPyPiRepoSimpleIndexUrl(String pypiRepoUrl) throws URISyntaxException {
-        URIBuilder pypiRepoUriBuilder = new URIBuilder(StringUtils.removeEnd(pypiRepoUrl, "/"));
-        List<String> repoUriPathSegments = pypiRepoUriBuilder.getPathSegments();
-        String lastPathSegment = CollectionUtils.isNotEmpty(repoUriPathSegments)
-                ? repoUriPathSegments.get(repoUriPathSegments.size() - 1)
-                : null;
-        if (!installDependenciesConfigurations.getPypiSimpleSuffix().equals(lastPathSegment)) {
-            // If the URL has no path, an unmodifiable Collections.emptyList() is returned,
-            // so wrap in an ArrayList to enable later modifications
-            repoUriPathSegments = new ArrayList<>(repoUriPathSegments);
-            repoUriPathSegments.add(installDependenciesConfigurations.getPypiSimpleSuffix());
-            pypiRepoUriBuilder.setPathSegments(repoUriPathSegments);
-        }
-
-        return StringUtils.appendIfMissing(pypiRepoUriBuilder.build().toString(), "/");
-    }
-
     protected void processManagedDependencyMismatches() {
-        if (!installDependenciesConfigurations.getManagedDependencies().isEmpty()) {
+        if (!installDependenciesMojo.getManagedDependencies().isEmpty()) {
             Map<String, TomlReplacementTuple> replacements = new HashMap<>();
 
             try (FileConfig pyProjectConfig = FileConfig.of(getPyProjectTomlFile())) {
@@ -230,7 +132,7 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
                 }
 
                 // Log replacements, if appropriate:
-                if (installDependenciesConfigurations.failOnManagedDependenciesMismatches() || !installDependenciesConfigurations.updateManagedDependenciesWhenFound()) {
+                if (installDependenciesMojo.failOnManagedDependenciesMismatches() || !installDependenciesMojo.updateManagedDependenciesWhenFound()) {
                     for (TomlReplacementTuple replacement : replacements.values()) {
                         logPackageMismatch(replacement.getPackageName(), replacement.getOriginalOperatorAndVersion(),
                                 replacement.getUpdatedOperatorAndVersion());
@@ -249,7 +151,7 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
         if (toolPoetryDependencies.isPresent()) {
             Config foundDependencies = toolPoetryDependencies.get();
             Map<String, Object> dependencyMap = foundDependencies.valueMap();
-            for (PackageDefinition def : installDependenciesConfigurations.getManagedDependencies()) {
+            for (PackageDefinition def : installDependenciesMojo.getManagedDependencies()) {
                 String packageName = def.getPackageName();
                 if (dependencyMap.containsKey(packageName)) {
                     Object packageRhs = dependencyMap.get(packageName);
@@ -272,7 +174,7 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
 
         String packageName = def.getPackageName();
 
-        if (installDependenciesConfigurations.overridePackageVersion() && updatedOperatorAndVersion.contains(SNAPSHOT)) {
+        if (installDependenciesMojo.overridePackageVersion() && updatedOperatorAndVersion.contains(SNAPSHOT)) {
             updatedOperatorAndVersion = replaceSnapshotWithWildcard(updatedOperatorAndVersion);
         }
 
@@ -291,8 +193,8 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
 
     protected void performPendingDependencyReplacements(Map<String, TomlReplacementTuple> replacements) {
         if (MapUtils.isNotEmpty(replacements)) {
-            if (installDependenciesConfigurations.failOnManagedDependenciesMismatches()) {
-                if (installDependenciesConfigurations.updateManagedDependenciesWhenFound()) {
+            if (installDependenciesMojo.failOnManagedDependenciesMismatches()) {
+                if (installDependenciesMojo.updateManagedDependenciesWhenFound()) {
                     log.warn("updateManagedDependenciesWhenFound=true will never be processed when failOnManagedDependenciesMismatches also equals true!");
                 }
 
@@ -300,7 +202,7 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
                         + "(see 'Package abc is not up to date with common project package definition guidance!` log messages above!");
             }
 
-            if (installDependenciesConfigurations.updateManagedDependenciesWhenFound()) {
+            if (installDependenciesMojo.updateManagedDependenciesWhenFound()) {
                 File pyProjectTomlFile = getPyProjectTomlFile();
                 String fileContent = StringUtils.EMPTY;
 
@@ -367,10 +269,6 @@ public class InstallDependenciesPoetry extends AbstractInstallDependencies {
 
         return operatorAndVersion;
 
-    }
-
-    protected static String replaceSnapshotWithWildcard(String pomVersion) {
-        return pomVersion.substring(0, pomVersion.indexOf(SNAPSHOT)) + ".*";
     }
 
     /**
